backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
boto3==1.35.0
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
requests==2.32.3
stripe==10.12.0
websockets==13.0
